requirements.txt
django-appconf==1.1.0
django-browser-reload==1.18.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-pint==0.7.3
django-vanilla-views==3.0.0
pillow==11.2.1
pipdeptree==2.26.1
requests==2.32.3
setuptools==65.5.0
sympy==1.13.3
Werkzeug==3.1.3
